[refactor] breakdown of proxy api classes
the proxy api had many classes in a single file, this commit simply break it down
