+++ lib/proxy_api/features.rb
@@ -1,0 +1,12 @@
+module ProxyAPI
+  class Features < Resource
+    def initialize args
+      @url  = args[:url] + "/features"
+      super args
+    end
+
+    def features
+      parse get
+    end
+  end
+end
